add load options

+++ cv/function.py
@@ -13,6 +13,9 @@ def dl_json(base_url: str, page: str) -> json:
         page: name of the page
     Return: json
     """
+    # Local only
+    #ssl._create_default_https_context = ssl._create_unverified_context
+
     json_url = base_url + page + ".json"
     with urllib.request.urlopen(json_url) as url:
         return_json = json.loads(url.read().decode())
@@ -21,15 +24,20 @@ def dl_json(base_url: str, page: str) -> json:
 
 
 @st.experimental_memo
-def dl_img(img_url: str, dest: str) -> Image:
+def dl_img(img_url: str, dest: str, force: bool = False) -> Image:
     """Get image from url
     Args:
         img_url: url of the image
         dest: dest to register the image
+        force: force dl image
     Return: Image
     """
+    # Local only
+    #ssl._create_default_https_context = ssl._create_unverified_context
+
     current_dir = os.path.dirname(os.path.abspath(__file__))
-    urllib.request.urlretrieve(img_url, "{}/".format(current_dir) + dest)
+    if not os.path.exists(str("{}/".format(current_dir) + dest)) or force:
+        urllib.request.urlretrieve(img_url, "{}/".format(current_dir) + dest)
     return_img = Image.open(str("{}/".format(current_dir) + dest))
 
     return return_img
